Poprawiono wygląd panelu stołu, dodano poprawne wyświetlanie pionków, dodano oblczanie średniej ważonej dla pionków drużyn, popraiwono skalowanie komponentów

+++ backend/Data/GameBoard.cs
@@ -9,8 +9,8 @@ namespace backend.Data
         public Game Game { get; set; } = null!;
         public int? GameProcessId{ get; set; }
         public GameProcess? GameProcess { get; set; }
-        public int PozX { get; set; }
-        public int PozY { get; set; }
+        public double PozX { get; set; }
+        public double PozY { get; set; }
         public int BoardId { get; set; }
         public Board Board { get; set; } = null!;
     }
